Implement faces of the n-cube and restriction stats
There is basic handling of faces of the n-cube, notably we can read and
write gaussoids as bitstrings in the same format as [^1] as well as a
representation of a face as a string over {0,1,*}.

The containment and restriction operators are implemented and the
test.p6 script implements a statistic which I became interested in this
afternoon: how often does each 3-gaussoid appear in the 3-dimensional
restrictions of 4-gaussoids?

The cnf4-list.txt data file is available from [^1] but included here
for convenience. The script takes 60 seconds to run at 100% CPU --
I can't hope to run it on cnf5-list.txt.

[^1] https://www.gaussoids.de/