Example of a set of squares without well-defined closure
The script exhibits an example (mostly done on paper, with closure
checks performed by the script) of a set of 2-faces for which multiple
non-isomorphic smallest extensions to a gaussoid exist.

In its current state, the script output isn't exciting. Try removing
some faces from @G and @H to see how it was useful.