Qnkpq-deg.p6: Implement degree formula
This took a very long time to figure out, even though it is "just"
counting the right things. The formula is only valid for n ≥ 2k,
which is likely just a remnant from my various failed attempts.
To lift that restriction should do nothing more than impose another
domain on m than 0 .. n-k, and that will depend on j.

In the Qnkpq/ directory, I include some adjacency lists for Q-graphs
because they take a long time to compute, and results of degree
computations by enumeration (Qnkpq-deg.p6 --enumerate) which can be
checked against the output of the --formula using Qnkpq-check.pl.
All of the data I gathered passes.