pairs.p6: list n-gaussoid pairs in opposite faces of (n+1)-gaussoids
The pairs.p6 program examines the pairs of n-gaussoids in the faces
0**...* and 1**...* of the (n+1)-cube, in all gaussoids on stdin.

This is quite feasible for 3-gaussoid pairs in the 4-gaussoids.
It won't work in its current state for n=5. After one night of running
the program I made a progress of < 3% (1.6 out of 60 million gaussoids).
I could throw more cores at it but it's 32°C these days even without
that. I have no immediate need for the statistics anyway.

For n=4, we have these statistics:

  # List each 3-gaussoid with how many different 3-gaussoids it appears
  $ ./pairs.p6 4 <cnf4-list.txt | sort -V | uniq | cut -f1 -d' ' | uniq -c
      5 000000
      9 000011
      9 001100
     10 011111
     10 101111
      9 110000
     10 110111
     10 111011
     10 111101
     10 111110
     11 111111

This shows that for 4-gaussoids at least, the full 3-gaussoid has
the greatest trouble being paired with other gaussoids. To see with
which gaussoids it can be paired:

  # List all pairs and how often they appear
  $ ./pairs.p6 4 <cnf4-list.txt | sort -V | uniq -c
      4 000000 000000
      1 000000 000011
      1 000000 001100
      1 000000 110000
      1 000000 111111
      1 000011 000000
      7 000011 000011
      ...

shows that exactly the six singletons are missing, which is expected.

It may be interesting to study the number of times each pair occurs
and how often these counts occur:

  $ ./pairs.p6 4 <cnf4-list.txt | sort -V | uniq -c | awk '{ print $1 }' | sort -V | uniq -c
     50 1
      1 4
      9 7
     24 8
      6 12
     12 20
      1 58