Add proof that completions.p6 is broken
The completion algorithm defied proof now that I'm in the last
stretches for my thesis and get around to examining it closer.
Turns out it did that because it's wrong. It does not produce
only the minimal gaussoid extensions but sometimes can also
print extensions which are not minimal.

Inspired by an earlier example where the full gaussoid in one
n-face of the (n+1)-cube brought easy-to-spot puzzle trouble,
I ran the script

  perl6 -e '.say for [X~] <0 1> xx 6' |
  while read other
    do ./pairs.p6 --compose 4 000000 $other
  done |
  while read H
    do echo $H >&2  # progress indicator
    ./completion.p6 --binary 4 $H
    echo
  done >full-completions.log

of which the output full-completions.log is also contained
in this commit. For the very first example, where it puts two
full 3-gaussoids into opposing n-faces, the problem surfaces
already. Some minimal extensions are printed and in addition
the full (n+1)-gaussoid is reported as an extension, but that
is clearly not minimal if non-full extensions exist.
(I'm genuinely surprised that this pair has multiple gaussoid
closures which lie below the full gaussoid.)

Thus, the algorithm needs a space- and memory-costly post-
processing step to be implemented soon.