completions.p6: add post-processing step to filter closures
Now completions.p6 does what it is meant to: report only gaussoid
closures, i.e. minimal extensions (which I confusingly used to call
completions). The correctness proof is now trivial because of that
costly last filtering step and the algorithm is not an achievement
at all -- at least without a good upper bound on the number of
extensions that the &extensions routine it will report. Maybe it
performs well in reducing the space of extensions to consider for
minimality, but that question sounds hard to approach.

The older examples outlines in the sources of completion.p6 did
not change.

I delete the full-completions.log. It isn't relevant anymore and
still in the git history for future reference.