Add tool to get a maximal independent subset from a Qnkpq.p6 adjacency list
This sort of complements tikz-Qnkpq.p6 to highlight an independent set
in such a graph. It is randomised, so you have to run it a few times
to get not only a maximal one but one with big cardinality, too.

It should be integrated into the tikz printer script but for now I just
did that by hand.